pyproject.toml
[project]
name = "tripletex-task"
version = "0.1.0"
description = "AI Accounting Agent for Tripletex - NM i AI 2026"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic[vertex]>=0.86.0",
    "fastapi>=0.115.0",
    "google-auth>=2.49.1",
    "google-cloud-aiplatform>=1.141.0",
    "google-genai>=1.68.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "numpy>=1.26.0",
    "pydantic-ai-slim[anthropic,logfire]>=1.70.0",
    "python-dotenv>=1.2.2",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=3.0.0",
    "uvicorn[standard]>=0.27.1",
]
